+++ src/selectors/metric_with_falloff.rs
@@ -5,12 +5,13 @@ use parry3d::math::{Point, Vector};
 use mesh_graph::{Face, MeshGraph, Selection};
 
 use super::{
-    MeshSelector, WeightedSelection, faces_incident_to_vertices,
+    DistanceCalculator, FalloffFn, L2, MeshSelector, WeightedSelection, faces_incident_to_vertices,
     get_sphere_with_falloff_weight_callback,
 };
 
 /// Generates a selection of a mesh that is within a sphere with a falloff
-pub struct SphereWithFalloff {
+#[derive(Debug)]
+pub struct MetricWithFalloff<D: DistanceCalculator> {
     /// The radius of the sphere.
     pub radius: f32,
 
@@ -19,24 +20,29 @@ pub struct SphereWithFalloff {
     /// The way the influence decreases is controlled by `falloff_func`.
     pub falloff: f32,
 
+    /// The metric squared used to calculate the distance between the input position and the vertices.
+    pub metric_squared: D,
+
     /// The falloff function used to calculate the weight of the selection.
     /// It receives values from 0.0 to 1.0 and has to return a value in the same range.
     /// Simply returning the input value is a linear falloff.
-    pub falloff_func: fn(f32) -> f32,
+    pub falloff_func: FalloffFn,
 }
 
-impl SphereWithFalloff {
+impl MetricWithFalloff<L2> {
+    /// Creates a new `MetricWithFalloff` selector with a sphere metric (normal L2 distance).
     #[inline]
-    pub fn new(radius: f32, falloff: f32, falloff_func: fn(f32) -> f32) -> Self {
+    pub fn sphere(radius: f32, falloff: f32, falloff_func: FalloffFn) -> Self {
         Self {
             radius,
             falloff,
+            metric_squared: L2,
             falloff_func,
         }
     }
 }
 
-impl MeshSelector for SphereWithFalloff {
+impl<D: DistanceCalculator + Copy + 'static> MeshSelector for MetricWithFalloff<D> {
     fn select(
         &self,
         mesh_graph: &MeshGraph,
@@ -62,7 +68,9 @@ impl MeshSelector for SphereWithFalloff {
         let max_dist_sqr = sum * sum;
 
         for vertex_id in potential_selection.resolve_to_vertices(mesh_graph) {
-            let distance = mesh_graph.positions[vertex_id].distance_squared(input_pos);
+            let distance = self
+                .metric_squared
+                .distance_squared(mesh_graph.positions[vertex_id], input_pos);
 
             if distance <= max_dist_sqr {
                 vertices.insert(vertex_id);
@@ -79,6 +87,7 @@ impl MeshSelector for SphereWithFalloff {
                 self.radius,
                 self.falloff,
                 self.falloff_func,
+                self.metric_squared,
             ),
         }
     }