+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2025-06-01
+
+- Made selector metric configurable
+- Smooth deformation is 90% less sensitive
+
 ## [0.2.0] - 2025-05-29
 
 - Made selector fields public